feat(handlers): update all menu handlers
Add check for user block status for early exit
Add `return void` for early `answerCallbackQuery` calls
In some cases, move final `answerCallbackQuery` after menu update

`closeMenuHandler`:
- Add `try/catch` for message delete
`dynamicListHandler`:
- Fix type for `range.text(...)` context
`fingerprintHandler`: Convert to arrow function
`outdatedHandler`:
- Add variable state for `catch` block
- Move `answerCallbackQuery` to `finally` block (Output depends on outcome of session and error state check)

+++ src/handlers/menu/outdatedHandler.ts
@@ -1,24 +1,35 @@
 import { MenuBotContext } from "@/src/types/bot.ts";
 import { locale } from "@/src/constants/locale.ts";
+import { isBotBlockedByUser } from "@/src/helpers/api.ts";
 
-const { outdated, failedToUpdate } = locale.frontend.menu;
+const { botBlocked, menu: { outdated, failedToUpdate } } = locale.frontend;
 
 export async function outdatedHandler(ctx: MenuBotContext) {
+    if (await isBotBlockedByUser(ctx)) {
+        return void await ctx.answerCallbackQuery(botBlocked);
+    }
+
+    let isFailedToUpdate = false;
+
     try {
         if (!ctx.session.currentFavorites) {
             await ctx.deleteMessage();
-            await ctx.answerCallbackQuery();
         } else {
             ctx.menu.update();
-            await ctx.answerCallbackQuery(outdated);
         }
     } catch (error: unknown) {
+        isFailedToUpdate = true;
         console.error(
             `Failed to run outdatedHandler: ${(error as Error).message}`,
         );
         await ctx.reply(failedToUpdate, {
             reply_to_message_id: ctx.message?.message_id,
         });
-        await ctx.answerCallbackQuery();
+    } finally {
+        await ctx.answerCallbackQuery(
+            ctx.session.currentFavorites && !isFailedToUpdate
+                ? outdated
+                : undefined,
+        );
     }
 }
